drivers: npmx: shell: move charger code to the separate file
