+++ drivers/npmx/CMakeLists.txt
@@ -49,6 +49,7 @@ if(CONFIG_NPMX_SHELL)
 
     zephyr_library_sources(shell/adc.c)
     zephyr_library_sources(shell/buck.c)
+    zephyr_library_sources(shell/charger.c)
 endif()
 
 # Get npmx include file
